Beginning of Arm Subsystem (#27)
* Beginning of Arm Subsystem

* Added motors and RobotMap file

I can guarantee with absolute 100% certainty that this code will NOT work. The imports aren't implemented and i'm not sure if I did the motors correctly.

The reason the imports are so bad is that I was unable to apply the gradle configurations to the project (likely issues with the school wifi) so I'll probably attempt to install the gradle configuration later. 

I based the motors off of the 2018 motor styles (a motor[] for the arm, and two individual motor variables for the two cargograbbers) so I'm not entirely sure if they work.

The RobotMap.java is (obviously) loaded with a bunch of bogus numbers.

* Added motors and RobotMap file

I can guarantee with absolute 100% certainty that this code will NOT work. The imports aren't implemented and i'm not sure if I did the motors correctly.

The reason the imports are so bad is that I was unable to apply the gradle configurations to the project (likely issues with the school wifi) so I'll probably attempt to install the gradle configuration later. 

I based the motors off of the 2018 motor styles (a motor[] for the arm, and two individual motor variables for the two cargograbbers) so I'm not entirely sure if they work.

The RobotMap.java is (obviously) loaded with a bunch of bogus numbers.

* Added subsystem to manager

(see title)

* Bug fix

Fixed "No newline at end of file" thing.

* Fixed imports

Fixed all imports (aka fixed gradle configuration so I could fix all the errors that were happening)

* Added armsubsystem to subsystemmanager

see title

* Removed unnecessary comment

Removed unnecessary comment

* Import fixes

Some import fixes

// meow

* Removed unused import

Removed unused import
(meow!)

* Updated common module

Updated common module

* Fix RobotMap Values

They're still bogus, except now they're not conflicting with the other (non-bogus) values.