Vision Code (#47)
* Add PnP Code

Implemented PnP (Point-in-Perspective) code. This is used to find how far away the robot is from the target. The code is real ugly but this commit is just for backup purposes.

* Cleaned up code, added new features

Cleaned up the code so it'll now work with any number of image points, so we can do PnP on the hatch or whatever. Implemented some fail safes as well so we won't crash if we get bad data. Updated the Limelight driver, fixed our sorting algorithm for the image points. The sorting algorithm for the image points should be updated if a better method is found.

* Documentation and other Small Changes

Added documentation to the pipeline and ImagePoint classes, and some small changes here and there. Tested and works.