Mount Vernon Changes (#53)
* Add PnP Code

Implemented PnP (Point-in-Perspective) code. This is used to find how far away the robot is from the target. The code is real ugly but this commit is just for backup purposes.

* Cleaned up code, added new features

Cleaned up the code so it'll now work with any number of image points, so we can do PnP on the hatch or whatever. Implemented some fail safes as well so we won't crash if we get bad data. Updated the Limelight driver, fixed our sorting algorithm for the image points. The sorting algorithm for the image points should be updated if a better method is found.

* Documentation and other Small Changes

Added documentation to the pipeline and ImagePoint classes, and some small changes here and there. Tested and works.

* bEGINning of Vision Coprocessor networking

* Adds networking code for periodic state updates for the vision coprocessor

* Adds control packet networking

* Updates submodule

* "Optimized" sorting algorithm, added calculations for angle to target

"Optimized" the sorting algorithm in the sense that I got rid of translatedPoint. If you don't know what that was then good, it just made the code uglier. I additionally added calculations to get our angle to the target. I do want to test this code a bit more, because initial testing wasn't entirely satisfactory, but it is stable and ready to be merged if it needs to be.

* Updates GradleRIO to version 2019.3.2

* Changes hatch placer to new mechanism

* Changes

* Practice field

* Changes made at practice field

* More changes

* Correctly zeroes gyroscope at beginning of auto

* Makes primary driver intake control override secondary driver

* Adds vision targeting

* Adds camera stream switcher which should lower bandwidth usage

* Makes the secondary controller vibrate when a hatch is gathered from the floor

* Makes arm go up after climbing automatically

* Rezeros cargo arm

* Gracefully aborts climb

* Makes the hatch floor grabber always intake on the intake sequence

* Rezeros drivetrain

* Updates submodule